+++ theboys/funcExtras.h
@@ -7,6 +7,6 @@
 int aleat(int min, int max);
 
 // retorna a distancia cartesiana entre dois pontos de um plano cartesiano
-int distancia_cartesiana(struct localizacao p1, struct localizacao p2);
+double distancia_cartesiana(struct localizacao p1, struct localizacao p2);
 
 #endif
